demos/text-ocr/pyproject.toml
[project]
name = "portfolio-demo-ocr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.30",
    "onnxruntime>=1.17",
    "numpy>=1.26",
    "Pillow>=10.0",
    "streamlit-drawable-canvas>=0.9",
]

[dependency-groups]
export = [
    "torch>=2.1",
    "onnx>=1.15",
    "onnxscript>=0.6.2",
]
dev = ["pytest>=8.0"]
